fix issues

+++ Models/PaymentGroup/GetPaymentGroupsRequest.cs
@@ -2,8 +2,8 @@ namespace MAS.Payments.Models
 {
     public class GetPaymentGroupsRequest
     {
-        public int? Month { get; set; }
+        public byte? Month { get; set; }
 
-        public int? Year { get; set; }
+        public short? Year { get; set; }
     }
 }
